chore(plc4x): Minor updates to plc4c plc4x driver

+++ plc4c/drivers/plc4x/src/driver_plc4x_packets.c
@@ -61,7 +61,7 @@ int16_t plc4c_driver_plc4x_select_message_function(uint8_t* buffer_data,
                                                    uint16_t buffer_length) {
   if (buffer_length >= 3) {
     uint16_t length = ((uint16_t) *(buffer_data + 1)) << 8 | ((uint16_t) *(buffer_data + 2));
-    return length;
+    return (int16_t) length;
   }
 
   // In all other cases, we'll just have to wait for the next time.
